chore: implements business logic

+++ src/main/java/com/github/emeps/cars/services/VehicleService.java
@@ -1,10 +1,13 @@
 package com.github.emeps.cars.services;
 
 import com.github.emeps.cars.DTOS.CreateVehicleData;
+import com.github.emeps.cars.DTOS.UpdateVehicleData;
 import com.github.emeps.cars.models.Vehicle;
+import com.github.emeps.cars.pagination.Page;
 import com.github.emeps.cars.repositories.ModelRepository;
 import com.github.emeps.cars.repositories.VehicleRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -34,12 +37,28 @@ public class VehicleService {
 
         return vehicle;
     }
-    public Vehicle update(Vehicle vehicle, Long id) {
-        return null;
+    public Vehicle update(UpdateVehicleData data, Long id) {
+
+        var vehicle = vehicleRepository.findById(id)
+                .orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"vehicle not found"));
+
+        var model = modelRepository.findById(data.getModelId())
+                .orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"model not found"));
+        vehicle.setKilometers(data.getKilometers());
+        vehicle.setColor(data.getColor());
+        vehicle.setDescription(data.getDescription());
+        vehicle.setYear(data.getYear());
+        vehicle.setModel(model);
+
+        vehicleRepository.save(vehicle);
+
+        return vehicle;
+
     }
 
-    public List<Vehicle> list() {
-        return  vehicleRepository.findAll();
+    public Page<Vehicle> list(Pageable pageable) {
+        var page =  vehicleRepository.findAll(pageable);
+        return  new Page<>(page);
     }
     public Vehicle show(Long id){
         return vehicleRepository.findById(id).orElseThrow(
@@ -48,8 +67,10 @@ public class VehicleService {
     }
 
     public void deleteById(Long id){
+        if (!vehicleRepository.existsById(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "vehicle not found");
+        }
         vehicleRepository.deleteById(id);
+        return;
     }
-
-
 }
